Use int64 for stream_id

+++ federationtypes.go
@@ -221,7 +221,7 @@ type respStateFields struct {
 // https://matrix.org/docs/spec/server_server/latest#get-matrix-federation-v1-user-devices-userid
 type RespUserDevices struct {
 	UserID         string           `json:"user_id"`
-	StreamID       int              `json:"stream_id"`
+	StreamID       int64            `json:"stream_id"`
 	Devices        []RespUserDevice `json:"devices"`
 	MasterKey      *CrossSigningKey `json:"master_key"`
 	SelfSigningKey *CrossSigningKey `json:"self_signing_key"`
